refactor(shared): split track-meta into facts, geometry, keys, join, verification
