+++ test/track-segment-generate.test.ts
@@ -17,7 +17,7 @@ import {
   writableAlignments,
 } from "../shared/track-segment-generate";
 import { loadTrackFacts, loadTrackGeometry } from "../shared/track-data";
-import type { TrackGeometry } from "../shared/track-meta";
+import type { TrackGeometry } from "../shared/track-geometry";
 import { officialTurnCount, validateFacts } from "../shared/track-segment-align";
 import { loadDetectHints } from "../shared/track-detect-hints";
 import { KNOWN_ALIGNMENT_GAPS, KNOWN_FUZZY_ALIGNMENTS } from "./helpers/track-known-gaps";
